Move sts-mock logger into common folder
