+++ sts-mock/src/functions/services/logging/README.md
