+++ sts-mock/src/functions/services/logging/commonRegisteredLogs.ts
