+++ sts-mock/src/functions/services/logging/logger.ts
