+++ sts-mock/src/functions/services/logging/tests/mockLoggingAdapter.ts
