+++ sts-mock/src/functions/services/logging/types.ts
