+++ sts-mock/src/functions/sts-mock/token/handlerDependencies.ts
@@ -1,6 +1,6 @@
 import { Logger as PowertoolsLogger } from "@aws-lambda-powertools/logger";
 import { MessageName, registeredLogs } from "./registeredLogs";
-import { Logger } from "../services/logging/logger";
+import { Logger } from "../../services/logging/logger";
 import {
   IValidateServiceTokenRequest,
   validateServiceTokenRequest,
