+++ sts-mock/src/functions/sts-mock/token/registeredLogs.ts
@@ -1,8 +1,8 @@
 import {
   CommonMessageNames,
   commonMessages,
-} from "../services/logging/commonRegisteredLogs";
-import { RegisteredLogMessages } from "../services/logging/types";
+} from "../../services/logging/commonRegisteredLogs";
+import { RegisteredLogMessages } from "../../services/logging/types";
 
 export type MessageName =
   | "INVALID_REQUEST"
