+++ sts-mock/src/functions/sts-mock/token/tests/tokenHandler.test.ts
@@ -1,5 +1,5 @@
-import { Logger } from "../../services/logging/logger";
-import { MockLoggingAdapter } from "../../services/logging/tests/mockLoggingAdapter";
+import { Logger } from "../../../services/logging/logger";
+import { MockLoggingAdapter } from "../../../services/logging/tests/mockLoggingAdapter";
 import { buildLambdaContext } from "../../../testUtils/mockContext";
 import { MessageName, registeredLogs } from "../registeredLogs";
 import { lambdaHandlerConstructor } from "../tokenHandler";
